pyproject.toml
[project]
name = "ai-assessment-platform"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.13.0",
    "diffusers>=0.37.0",
    "elevenlabs>=2.39.0",
    "ipykernel>=7.2.0",
    "langchain>=1.2.12",
    "langchain-classic>=1.0.3",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.2",
    "langchain-openai>=1.1.11",
    "pillow>=12.1.1",
    "python-dotenv>=1.2.2",
    "sarvamai>=0.1.27",
    "scipy>=1.17.1",
    "sounddevice>=0.5.5",
    "streamlit>=1.55.0",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "transformers>=5.3.0",
]
